consolidate LDAP and LDAPHierarchical auth
Thanks to the previously merged DS-1180, the LDAPHierarchicalAuthentication
method gaind the capabilities of LDAPAuthentication. This commit renames
LDAPHierarchicalAuthentication to LDAPAuthentication and removes the
original LDAPAuthentication.
