+++ dspace/config/modules/authentication-ldap.cfg
@@ -6,8 +6,7 @@
 #---------------------------------------------------------------#
 #
 # In order to enable LDAP Authentication, you must first ensure the
-# 'org.dspace.authenticate.LDAPAuthentication' OR 
-# 'org.dspace.authenticate.LDAPHierarchicalAuthentication'
+# 'org.dspace.authenticate.LDAPAuthentication'
 # class is added to the list of enabled AuthenticationMethods in 'authenticate.cfg'.  
 # See 'authenticate.cfg' for more info.
 #
@@ -112,10 +111,9 @@ autoregister = true
 ##### Hierarchical LDAP Settings #####
 
 # If your users are spread out across a hierarchical tree on your
-# LDAP server, you will need to use the following stackable authentication
-# class:
-#  plugin.sequence.org.dspace.authenticate.AuthenticationMethod = \
-#        org.dspace.authenticate.LDAPHierarchicalAuthentication
+# LDAP server, you will need to specify the username and password of
+# a user who is allowed to search the tree to find the full DN of
+# the user who is logging in.
 #
 # You can optionally specify the search scope. If anonymous access is not
 # enabled on your LDAP server, you will need to specify the full DN and
