+++ dspace/config/modules/authentication.cfg
@@ -17,7 +17,6 @@
 #        Configuration file: authentication-shibboleth.cfg
 #  * LDAP Authentication
 #        LDAP Plugin class: org.dspace.authenticate.LDAPAuthentication
-#        Hierarchical LDAP Plugin class: org.dspace.authenticate.LDAPHierarchicalAuthentication
 #        Configuration file: authentication-ldap.cfg
 #  * IP Address Authentication
 #        Plugin class: org.dspace.authenticate.IPAuthentication
